+++ tests/Nelmio/Alice/Parser/RuntimeCacheParserTest.php
@@ -19,7 +19,14 @@ use Prophecy\Argument;
  */
 class RuntimeCacheParserTest extends \PHPUnit_Framework_TestCase
 {
-    public function test_is_a_parser()
+    private static $dir;
+
+    public function setUp()
+    {
+        self::$dir = __DIR__.'/File/Cache';
+    }
+
+    public function testIsAParser()
     {
         $this->assertTrue(is_a(RuntimeCacheParser::class, ParserInterface::class, true));
     }
@@ -27,7 +34,7 @@ class RuntimeCacheParserTest extends \PHPUnit_Framework_TestCase
     /**
      * @dataProvider provideParsableFile
      */
-    public function test_can_parse_file(string $file)
+    public function testCanParseFile(string $file)
     {
         $expected = [new \stdClass()];
 
@@ -41,13 +48,13 @@ class RuntimeCacheParserTest extends \PHPUnit_Framework_TestCase
 
         $this->assertSame($expected, $actual);
 
-        // as the parser cache the result, run a second time
+        // As the parser cache the results, parsing each file does not re-trigger a parse call
         $actual = $parser->parse($file);
 
         $this->assertSame($expected, $actual);
     }
 
-    public function test_cache_parse_result()
+    public function testCacheParseResult()
     {
         $realFilePath = __FILE__;
         $file1 = $realFilePath;
@@ -69,7 +76,7 @@ class RuntimeCacheParserTest extends \PHPUnit_Framework_TestCase
         $decoratedParserProphecy->parse(Argument::any())->shouldHaveBeenCalledTimes(1);
     }
 
-    public function test_dont_cache_parse_result_with_no_absolute_path_could_be_retrieved()
+    public function testDontCacheParseResultIfNoAbsolutePathCouldBeRetrieved()
     {
         $file = 'https://example.com/script.php';
         $expected = [new \stdClass()];
@@ -87,6 +94,81 @@ class RuntimeCacheParserTest extends \PHPUnit_Framework_TestCase
         $this->assertSame($actual1, $actual2);
 
         $decoratedParserProphecy->parse(Argument::any())->shouldHaveBeenCalledTimes(2);
+    }
+
+    public function testProcessIncludesAndCacheTheResultOfEachIncludedFile()
+    {
+        $mainFile = self::$dir.'/main.yml';   // needs to be a real file to be cached
+        $parsedMainFileContent = [
+            'include' => [
+                'file1.yml',
+                'another_level/file2.yml',
+            ],
+            'Nelmio\Alice\Model\User' => [
+                'user_main' => [],
+            ],
+        ];
+        $parsedFile1Content = [
+            'Nelmio\Alice\Model\User' => [
+                'user_file1' => [],
+            ],
+        ];
+        $parsedFile2Content = [
+            'include' => [
+                self::$dir.'/file3.yml',
+            ],
+            'Nelmio\Alice\Model\User' => [
+                'user_file2' => [],
+            ],
+        ];
+        $parsedFile3Content = [
+            'Nelmio\Alice\Model\User' => [
+                'user_file3' => [],
+            ],
+        ];
+        $expected = [
+            'Nelmio\Alice\Model\User' => [
+                'user_file3' => [],
+                'user_file2' => [],
+                'user_file1' => [],
+                'user_main' => [],
+            ],
+        ];
+        $expectedFile2 = [
+            'Nelmio\Alice\Model\User' => [
+                'user_file3' => [],
+                'user_file2' => [],
+            ],
+        ];
+
+        $decoratedParserProphecy = $this->prophesize(ParserInterface::class);
+        $decoratedParserProphecy->parse($mainFile)->willReturn($parsedMainFileContent);
+        $decoratedParserProphecy->parse(Argument::containingString('file1.yml'))->willReturn($parsedFile1Content);
+        $decoratedParserProphecy->parse(Argument::containingString('file2.yml'))->willReturn($parsedFile2Content);
+        $decoratedParserProphecy->parse(Argument::containingString('file3.yml'))->willReturn($parsedFile3Content);
+        /* @var ParserInterface $decoratedParser */
+        $decoratedParser = $decoratedParserProphecy->reveal();
+
+        $parser = new RuntimeCacheParser($decoratedParser);
+        $actual = $parser->parse($mainFile);
+
+        $this->assertSame($expected, $actual);
+
+        // As the parser cache the results, parsing each file does not re-trigger a parse call
+        $actual = $parser->parse($mainFile);
+        $actualFile1 = $parser->parse('file1.yml');
+        $actualFile2 = $parser->parse('file2.yml');
+        $actualFile3 = $parser->parse('file3.yml');
+
+        $this->assertSame($expected, $actual);
+        $this->assertSame($parsedFile1Content, $actualFile1);
+        $this->assertSame($expectedFile2, $actualFile2);
+        $this->assertSame($parsedFile3Content, $actualFile3);
+
+        $decoratedParserProphecy->parse($mainFile)->shouldHaveBeenCalledTimes(1);
+        $decoratedParserProphecy->parse('file1.yml')->shouldHaveBeenCalledTimes(1);
+        $decoratedParserProphecy->parse('file2.yml')->shouldHaveBeenCalledTimes(1);
+        $decoratedParserProphecy->parse('file3.yml')->shouldHaveBeenCalledTimes(1);
     }
 
     public function provideParsableFile()